apps/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2

# Image processing
pillow==10.1.0
numpy==1.24.3
rembg==2.0.50

# External APIs
replicate==0.22.0
requests==2.31.0

# Production dependencies
gunicorn==21.2.0
psutil==5.9.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring (optional)
prometheus-client==0.19.0
